# Improve the performance of `partition` for lazy and strict bytestrings