--- a/bench/bench-bytestring.cabal
+++ b/bench/bench-bytestring.cabal
@@ -42,6 +42,7 @@ executable bench-bytestring-builder
                    -- we require bytestring due to benchmarking against
                    -- blaze-textual, which uses blaze-builder
                    , bytestring    >= 0.9
+                   , random
 
   -- cabal complains about ../ dirs. However, this is better than symlinks,
   -- which probably don't work on windows.