@gregorycollins 2015-09-29
Looks good, please send some benchmark results along. @tibbe