@SeanRBurton 2015-09-30
Benchmark results:

New: 

```
benchmarking partition/strict/mostlyTrueFast
mean: 192.5830 us, lb 191.0881 us, ub 194.6811 us, ci 0.950
std dev: 8.992629 us, lb 6.969597 us, ub 11.20351 us, ci 0.950
found 43 outliers among 100 samples (43.0%)
  22 (22.0%) low severe
  21 (21.0%) high severe
variance introduced by outliers: 44.484%
variance is moderately inflated by outliers

benchmarking partition/strict/mostlyFalseFast
mean: 181.7812 us, lb 181.3666 us, ub 182.2942 us, ci 0.950
std dev: 2.370496 us, lb 1.996720 us, ub 3.097616 us, ci 0.950
found 2 outliers among 100 samples (2.0%)
  1 (1.0%) high severe
variance introduced by outliers: 6.550%
variance is slightly inflated by outliers

benchmarking partition/strict/balancedFast
mean: 258.3463 us, lb 257.7349 us, ub 259.2614 us, ci 0.950
std dev: 3.782602 us, lb 2.832685 us, ub 6.178624 us, ci 0.950
found 2 outliers among 100 samples (2.0%)
  1 (1.0%) high severe
variance introduced by outliers: 7.534%
variance is slightly inflated by outliers

benchmarking partition/strict/mostlyTrueSlow
mean: 5.266672 ms, lb 5.255096 ms, ub 5.283814 ms, ci 0.950
std dev: 71.40990 us, lb 52.40680 us, ub 102.4179 us, ci 0.950
found 4 outliers among 100 samples (4.0%)
  2 (2.0%) high mild
  2 (2.0%) high severe
variance introduced by outliers: 6.584%
variance is slightly inflated by outliers

benchmarking partition/strict/mostlyFalseSlow
mean: 5.336793 ms, lb 5.320993 ms, ub 5.359226 ms, ci 0.950
std dev: 95.21262 us, lb 72.59695 us, ub 126.2094 us, ci 0.950
found 7 outliers among 100 samples (7.0%)
  6 (6.0%) high severe
variance introduced by outliers: 10.406%
variance is moderately inflated by outliers

benchmarking partition/strict/balancedSlow
mean: 5.415187 ms, lb 5.404451 ms, ub 5.431784 ms, ci 0.950
std dev: 67.38540 us, lb 48.52077 us, ub 102.8109 us, ci 0.950

benchmarking partition/lazy/mostlyTrueFast
mean: 206.3493 us, lb 202.1746 us, ub 212.4313 us, ci 0.950
std dev: 25.44986 us, lb 19.05630 us, ub 33.55232 us, ci 0.950
found 8 outliers among 100 samples (8.0%)
  7 (7.0%) high severe
variance introduced by outliers: 85.240%
variance is severely inflated by outliers

benchmarking partition/lazy/mostlyFalseFast
mean: 184.7712 us, lb 184.3403 us, ub 185.4353 us, ci 0.950
std dev: 2.708211 us, lb 1.988782 us, ub 4.586251 us, ci 0.950
found 2 outliers among 100 samples (2.0%)
  1 (1.0%) high severe
variance introduced by outliers: 7.535%
variance is slightly inflated by outliers

benchmarking partition/lazy/balancedFast
mean: 261.9191 us, lb 261.1291 us, ub 263.4365 us, ci 0.950
std dev: 5.417403 us, lb 3.125183 us, ub 9.000554 us, ci 0.950
found 5 outliers among 100 samples (5.0%)
  2 (2.0%) high mild
  3 (3.0%) high severe
variance introduced by outliers: 13.284%
variance is moderately inflated by outliers

benchmarking partition/lazy/mostlyTrueSlow
mean: 5.344801 ms, lb 5.317903 ms, ub 5.401755 ms, ci 0.950
std dev: 191.6010 us, lb 99.19357 us, ub 328.6404 us, ci 0.950
found 10 outliers among 100 samples (10.0%)
  5 (5.0%) high mild
  5 (5.0%) high severe
variance introduced by outliers: 32.594%
variance is moderately inflated by outliers

benchmarking partition/lazy/mostlyFalseSlow
mean: 5.359943 ms, lb 5.342327 ms, ub 5.385180 ms, ci 0.950
std dev: 107.3547 us, lb 80.67685 us, ub 144.5866 us, ci 0.950
found 7 outliers among 100 samples (7.0%)
  2 (2.0%) high mild
  5 (5.0%) high severe
variance introduced by outliers: 13.237%
variance is moderately inflated by outliers

benchmarking partition/lazy/balancedSlow
mean: 5.475352 ms, lb 5.459919 ms, ub 5.498486 ms, ci 0.950
std dev: 95.48023 us, lb 71.76160 us, ub 141.1447 us, ci 0.950
found 8 outliers among 100 samples (8.0%)
  4 (4.0%) high mild
  4 (4.0%) high severe
variance introduced by outliers: 10.378%
variance is moderately inflated by outliers
```

Old:

```
benchmarking partition/strict/mostlyTrueFast
mean: 356.7592 us, lb 354.5205 us, ub 359.7466 us, ci 0.950
std dev: 13.14770 us, lb 10.43953 us, ub 16.66730 us, ci 0.950
found 11 outliers among 100 samples (11.0%)
  6 (6.0%) high mild
  5 (5.0%) high severe
variance introduced by outliers: 33.591%
variance is moderately inflated by outliers

benchmarking partition/strict/mostlyFalseFast
mean: 316.8055 us, lb 316.0418 us, ub 317.7841 us, ci 0.950
std dev: 4.422674 us, lb 3.600381 us, ub 5.621884 us, ci 0.950
found 5 outliers among 100 samples (5.0%)
  5 (5.0%) high mild
variance introduced by outliers: 6.607%
variance is slightly inflated by outliers

benchmarking partition/strict/balancedFast
mean: 439.2636 us, lb 436.9513 us, ub 442.2241 us, ci 0.950
std dev: 13.37241 us, lb 10.86467 us, ub 17.65873 us, ci 0.950
found 2 outliers among 100 samples (2.0%)
  1 (1.0%) high severe
variance introduced by outliers: 25.771%
variance is moderately inflated by outliers

benchmarking partition/strict/mostlyTrueSlow
mean: 10.82000 ms, lb 10.77080 ms, ub 10.90243 ms, ci 0.950
std dev: 316.3557 us, lb 215.0004 us, ub 527.7903 us, ci 0.950
found 3 outliers among 100 samples (3.0%)
  2 (2.0%) high severe
variance introduced by outliers: 23.858%
variance is moderately inflated by outliers

benchmarking partition/strict/mostlyFalseSlow
mean: 10.68479 ms, lb 10.66405 ms, ub 10.71633 ms, ci 0.950
std dev: 128.8148 us, lb 93.79337 us, ub 183.6947 us, ci 0.950

benchmarking partition/strict/balancedSlow
mean: 10.94958 ms, lb 10.84616 ms, ub 11.09129 ms, ci 0.950
std dev: 617.9129 us, lb 484.0755 us, ub 754.3261 us, ci 0.950
found 20 outliers among 100 samples (20.0%)
  4 (4.0%) high mild
  16 (16.0%) high severe
variance introduced by outliers: 54.461%
variance is severely inflated by outliers

benchmarking partition/lazy/mostlyTrueFast
mean: 359.2412 us, lb 357.2504 us, ub 362.0593 us, ci 0.950
std dev: 12.05215 us, lb 9.377301 us, ub 15.64474 us, ci 0.950
found 23 outliers among 100 samples (23.0%)
  23 (23.0%) high severe
variance introduced by outliers: 29.666%
variance is moderately inflated by outliers

benchmarking partition/lazy/mostlyFalseFast
mean: 322.1620 us, lb 321.0517 us, ub 324.1202 us, ci 0.950
std dev: 7.343042 us, lb 4.847093 us, ub 12.61059 us, ci 0.950
found 6 outliers among 100 samples (6.0%)
  4 (4.0%) high mild
  2 (2.0%) high severe
variance introduced by outliers: 16.131%
variance is moderately inflated by outliers

benchmarking partition/lazy/balancedFast
mean: 433.7163 us, lb 431.5402 us, ub 439.1223 us, ci 0.950
std dev: 16.45033 us, lb 7.236713 us, ub 33.47401 us, ci 0.950
found 10 outliers among 100 samples (10.0%)
  7 (7.0%) high mild
  3 (3.0%) high severe
variance introduced by outliers: 34.592%
variance is moderately inflated by outliers

benchmarking partition/lazy/mostlyTrueSlow
mean: 10.73677 ms, lb 10.71079 ms, ub 10.78405 ms, ci 0.950
std dev: 175.2876 us, lb 107.3593 us, ub 278.7758 us, ci 0.950
found 10 outliers among 100 samples (10.0%)
  7 (7.0%) high mild
  3 (3.0%) high severe
variance introduced by outliers: 9.414%
variance is slightly inflated by outliers

benchmarking partition/lazy/mostlyFalseSlow
mean: 10.74606 ms, lb 10.72162 ms, ub 10.79496 ms, ci 0.950
std dev: 169.5766 us, lb 97.13227 us, ub 298.2051 us, ci 0.950
found 8 outliers among 100 samples (8.0%)
  3 (3.0%) high mild
  5 (5.0%) high severe
variance introduced by outliers: 8.492%
variance is slightly inflated by outliers

benchmarking partition/lazy/balancedSlow
mean: 10.79299 ms, lb 10.75831 ms, ub 10.85812 ms, ci 0.950
std dev: 235.4542 us, lb 140.2205 us, ub 379.1693 us, ci 0.950
found 8 outliers among 100 samples (8.0%)
  2 (2.0%) high mild
  6 (6.0%) high severe
variance introduced by outliers: 15.160%
variance is moderately inflated by outliers
```