@dcoutts 2015-09-30
Looks fine. Again, assuming our test coverage here is fine.